+++ server/src/main/java/luan/moonvs/repositories/RatingRepository.java
@@ -11,7 +11,10 @@ import java.util.List;
 import java.util.UUID;
 
 public interface RatingRepository extends JpaRepository<Rating, ContentAndUserId> {
+    @Deprecated
     List<Rating> getByIdRatingIdUser(UUID idUser);
+
+    List<Rating> getAllRatingsByIdRatingIdUser(UUID idUser);
 
     @Query("""
             SELECT AVG(r.ratingValue)