+++ server/src/main/java/luan/moonvs/services/RatingService.java
@@ -3,27 +3,33 @@ package luan.moonvs.services;
 import jakarta.persistence.EntityNotFoundException;
 import luan.moonvs.models.builders.RatingBuilder;
 import luan.moonvs.models.entities.ContentAndUserId;
+import luan.moonvs.models.entities.Profile;
 import luan.moonvs.models.entities.Rating;
 import luan.moonvs.models.entities.User;
 import luan.moonvs.models.requests.RateRequest;
 import luan.moonvs.models.requests.RatingRequest;
+import luan.moonvs.models.responses.Response;
 import luan.moonvs.repositories.RatingRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.UUID;
 
 @Service
 public class RatingService {
     private RatingRepository repository;
     private AccountService accountService;
+    private ProfileService profileService;
 
     @Autowired
-    public RatingService(RatingRepository repository, AccountService accountService) {
+    public RatingService(RatingRepository repository, AccountService accountService, ProfileService profileService) {
         this.repository = repository;
         this.accountService = accountService;
+        this.profileService = profileService;
     }
 
     @Deprecated
@@ -44,11 +50,12 @@ public class RatingService {
                 .body(rating);
     }
 
-    public ResponseEntity<Rating> addOrEditRating(int idContent, RateRequest rateRequest) {
-        User authUser = accountService.getAccount();
+    public Response<Rating> addOrEditRating(int idContent, RateRequest rateRequest) {
+        User user = accountService.getAccount();
+        ContentAndUserId idRating = new ContentAndUserId(user.getIdUser(), idContent);
         Rating rating;
         try {
-            rating = repository.getReferenceById(new ContentAndUserId(authUser.getIdUser(), idContent));
+            rating = repository.getReferenceById(idRating);
 
             rating = RatingBuilder.create(rating)
                     .addRating(rateRequest.ratingValue())
@@ -56,7 +63,7 @@ public class RatingService {
                     .build();
         } catch (EntityNotFoundException e) {
             rating = RatingBuilder.create()
-                    .withId(authUser.getIdUser(), idContent)
+                    .withId(user.getIdUser(), idContent)
                     .addRating(rateRequest.ratingValue())
                     .addCommentary(rateRequest.commentary())
                     .build();
@@ -64,34 +71,33 @@ public class RatingService {
 
         repository.save(rating);
 
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(rating);
+        return new Response<>(HttpStatus.OK, rating);
     }
 
-    public ResponseEntity<Rating> getUserRating(int idContent) {
-        User authUser = accountService.getAccount();
-        ContentAndUserId idRating = new ContentAndUserId(authUser.getIdUser(), idContent);
+    public Response<Rating> getUserRating(int idContent) {
+        final String NOT_FOUND = "It wasn't possible to find the rating";
+        User user = accountService.getAccount();
+        ContentAndUserId idRating = new ContentAndUserId(user.getIdUser(), idContent);
 
-        Rating rating = repository.getReferenceById(idRating);
+        try {
+            Rating rating = repository.getReferenceById(idRating);
 
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(rating);
+            return new Response<>(HttpStatus.OK, rating);
+        } catch (EntityNotFoundException e) {
+            return new Response<>(HttpStatus.NOT_FOUND, new Rating(), NOT_FOUND);
+        }
     }
 
-    public ResponseEntity<Float> getAvgRating(int idContent) {
+    public Response<Float> getAvgRating(int idContent) {
+        final String NOT_FOUND = "The content has no ratings";
         Float avgRating = repository.getAverageRatingByContent(idContent);
         if (avgRating == null)
-            return ResponseEntity
-                    .status(HttpStatus.NOT_FOUND)
-                    .build();
+            return new Response<>(HttpStatus.NOT_FOUND, Float.NaN, NOT_FOUND);
 
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(avgRating);
+        return new Response<>(HttpStatus.OK, avgRating);
     }
 
+    @Deprecated
     public ResponseEntity<?> getAllUserRatings() {
         User authUser = accountService.getAccount();
         try {
@@ -104,6 +110,22 @@ public class RatingService {
             return ResponseEntity
                     .status(HttpStatus.NOT_FOUND)
                     .body("Perfil não encontrado");
+        }
+    }
+
+    public Response<List<Rating>> getUserRatingList(UUID idUser) {
+        final String PROFILE_NOT_FOUND = "It wasn't possible to find the profile!",
+                     RATINGS_NOT_FOUND = "It wasn't possible to finde this profile's rating!";
+
+        Response<Profile> profile = profileService.getProfile(idUser);
+        if (!profile.status().is2xxSuccessful())
+            return new Response<>(HttpStatus.NOT_FOUND, Collections.emptyList(), PROFILE_NOT_FOUND);
+
+        try {
+            List<Rating> ratingList = repository.getAllRatingsByIdRatingIdUser(idUser);
+            return new Response<>(HttpStatus.OK, ratingList);
+        } catch (EntityNotFoundException e) {
+            return new Response<>(HttpStatus.NOT_FOUND, Collections.emptyList(), RATINGS_NOT_FOUND);
         }
     }
 }