+++ CHANGELOG.md
@@ -10,10 +10,15 @@
 - Backend exception 'IllegalIdException'.
 - Backend Account methods 'getAccount' and 'updateUsernameAndOrEmail'.
 - 'create()' methods on UserBuilder.
+- Copy Constructor on User entity.
 - 'create()' methods on ContentBuilder.
 - Backend Content methods 'search', 'databaseSearch', 'tmdbSearch', 'viewContent', 'viewDatabaseContent' and 'viewTmdbContent'.
 - SearchType enum.
 - Response DTO.
+- Backend Profile method 'getProfile'.
+- Backend Rating method 'getUserRatingList'.
+- Copy Constructor on Profile entity.
+- 'create()' methods on Profilebuilder.
 
 ## Changed
 
@@ -27,6 +32,8 @@
 - AccountController now performs basic validations.
 - Backend Account methods 'updatePassword' and 'deleteAccount'.
 - Backend Auth methods: 'login' and 'register'.
+- Backend Profile methods 'createProfile' and 'editProfile'.
+- Backend Rating methods 'addOrEditRating', 'getUserRating' and 'getAvgRating'.
 
 ## Deprecated
 
@@ -38,6 +45,9 @@
 - UserBuilder method 'withRegisterDto'.
 - Backend Auth methods: 'login' and 'register'.
 - Backend Content methods 'internalSearch', 'externalSearch', 'internalContent' and 'externalContent'
+- Backend Profile methods 'createProfile', 'seeProfile' and 'editProfile'.
+- Backend Rating method 'getAllUserRatings'.
+- ProfileBuilder constructors and methods 'createProfile' and 'editProfile'.
 
 ## Fixed
 
@@ -45,5 +55,6 @@
 - Fixed replacement of blank spaces on the search methods!
 - External links now open on a blank tab.
 - Fixed bug of titles in list of watch providers.
+- Fixed NullPointerException on ProfileBuilder method 'withBio'.
 
 [You can also access the full changelog](https://luanpozzobon.github.io/luanpozzobon_site/pages/projects/moonvs/moonvs.html)